feat: svg_to_pdf reuse conversion
