+++ node-graph/graphene-cli/src/export.rs
@@ -66,7 +66,7 @@ pub async fn export_document(
 		TaggedValue::RenderOutput(output) => match output.data {
 			RenderOutputType::Svg { svg, .. } => {
 				if file_type == FileType::Pdf {
-					let pdf_bytes = svg_to_pdf(&svg, width, height)?;
+					let pdf_bytes = rendering::svg_to_pdf::svg_to_pdf(&svg, width.map(|w| w as f32), height.map(|h| h as f32))?;
 					std::fs::write(&output_path, pdf_bytes)?;
 					log::info!("Exported PDF to: {}", output_path.display());
 				} else {
@@ -234,36 +234,4 @@ pub async fn export_gif(
 
 	log::info!("Exported GIF to: {}", output_path.display());
 	Ok(())
-}
-
-/// Convert an SVG string to PDF bytes using krilla and krilla-svg.
-fn svg_to_pdf(svg: &str, width: Option<u32>, height: Option<u32>) -> Result<Vec<u8>, String> {
-	use krilla::page::PageSettings;
-	use krilla_svg::{SurfaceExt, SvgSettings};
-
-	// Parse the SVG with usvg 0.45 (matching krilla-svg's expected version)
-	let options = usvg_045::Options::default();
-	let tree = usvg_045::Tree::from_str(svg, &options).map_err(|e| format!("Failed to parse SVG for PDF conversion: {e}"))?;
-
-	let (page_width, page_height) = match (width, height) {
-		(Some(w), Some(h)) => (w as f32, h as f32),
-		_ => (tree.size().width() as f32, tree.size().height() as f32),
-	};
-
-	// Create a krilla PDF document
-	let mut document = krilla::Document::new();
-
-	let page_settings = PageSettings::from_wh(page_width, page_height).ok_or_else(|| "Invalid page dimensions for PDF export".to_string())?;
-
-	let mut page = document.start_page_with(page_settings);
-	let mut surface = page.surface();
-
-	let krilla_size = krilla::geom::Size::from_wh(page_width, page_height).ok_or_else(|| "Invalid size for PDF export".to_string())?;
-	surface.draw_svg(&tree, krilla_size, SvgSettings::default());
-
-	surface.finish();
-	page.finish();
-
-	let pdf_bytes = document.finish().map_err(|e| format!("Failed to generate PDF: {e:?}"))?;
-	Ok(pdf_bytes)
 }
