+++ node-graph/libraries/rendering/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod convert_usvg_path;
+pub mod svg_to_pdf;
 pub mod render_ext;
 mod renderer;
 pub mod to_peniko;
